+++ lact-daemon/src/server/handler.rs
@@ -235,11 +235,7 @@ impl<'a> Handler {
         Ok(modes_table)
     }
 
-    pub async fn set_power_profile_mode(
-        &self,
-        id: &str,
-        index: Option<usize>,
-    ) -> anyhow::Result<()> {
+    pub async fn set_power_profile_mode(&self, id: &str, index: Option<u16>) -> anyhow::Result<()> {
         self.edit_gpu_config(id.to_owned(), |gpu_config| {
             gpu_config.power_profile_mode_index = index;
         })
